refactor(app): refactor RobotSettingsDashboard to simplify
Make RobotSettingButton into component to make RobotSettingsDashboard more maintainable
