+++ app/src/organisms/OnDeviceDisplay/RobotSettingsDashboard/DeviceReset.tsx
@@ -28,7 +28,7 @@ import { useDispatchApiRequest } from '../../../redux/robot-api'
 
 import type { Dispatch, State } from '../../../redux/types'
 import type { ResetConfigRequest } from '../../../redux/robot-admin/types'
-import type { SettingOption } from '../../../pages/OnDeviceDisplay/RobotSettingsDashboard'
+import type { SettingOption } from '../../../pages/OnDeviceDisplay/RobotSettingsDashboard/RobotSettingButton'
 
 interface LabelProps {
   isSelected?: boolean
