+++ app/src/organisms/OnDeviceDisplay/RobotSettingsDashboard/NetworkSettings/index.tsx
@@ -26,7 +26,7 @@ import { WifiConnectionDetails } from './WifiConnectionDetails'
 import type { IconName } from '@opentrons/components'
 import type { NetworkConnection } from '../../../../pages/OnDeviceDisplay/hooks'
 import type { ChipType } from '../../../../atoms/Chip'
-import type { SettingOption } from '../../../../pages/OnDeviceDisplay/RobotSettingsDashboard'
+import type { SettingOption } from '../../../../pages/OnDeviceDisplay/RobotSettingsDashboard/RobotSettingButton'
 
 export type ConnectionType = 'wifi' | 'ethernet' // TODO (kj: 04/05/2023) add 'usb' as options once implemented
