+++ app/src/organisms/OnDeviceDisplay/RobotSettingsDashboard/TextSize.tsx
@@ -19,7 +19,7 @@ import {
 
 import { StyledText } from '../../../atoms/text'
 
-import type { SettingOption } from '../../../pages/OnDeviceDisplay/RobotSettingsDashboard'
+import type { SettingOption } from '../../../pages/OnDeviceDisplay/RobotSettingsDashboard/RobotSettingButton'
 
 interface RectProps {
   isActive: boolean
